Update kernel-tests bug tracking URL
The kernel-tests scripts are somewhat disorganized because of the lack of an easy-to-use place to contribute upstream. The bug reported on Bugzilla was closed by a member of the Fedora kernel team because they said Bugzilla was not the appropriate place to report it, and instead encouraged posting to the mailing list, which I did. In the meanwhile, because of discussion in #fedora-python, it seems like rodrigc's repo is the best place for bug tracking as of now, without a central place upstream to manage the code.