[SCHEMA] Get "in-progress" statuses for Fedora from Bugzilla
The py3query script now gets bugs from Bugzilla, using a hard-coded
list of tracker bugs.
Bug status is put in the DB, and shown on the package page. In
summary listings, it's shown as a tooltip, and the bug icon is colored
accordingly.

If a package is not released yet, bug statuses other than NEW cause
it to be marked 'in-progress'. (This applies to CLOSED as well:
.)the package is only marked "released" when the actual RPMs with
py3 dependencies are seen.)

https://github.com/fedora-python/portingdb/issues/56