Fix pie charts in Jinja 2.9
In the new version of Jinja, it's impossible to assign to variables
in outer scopes -- and also, "for" loops are somehow scopes (see
[Jinja issue 641]).

This breaks SVG piecharts, which need the calculated angle from the
a segment to position the next one.

This will be fixed in [Jinja pull request 684], but currently it
needs a workaround.

[Jinja issue 641]: https://github.com/pallets/jinja/issues/641
[Jinja pull request 684]: https://github.com/pallets/jinja/pull/684