More polish for the history graph
I went to the Dash/Plotly workshop at PyCon CZ, and had a short
discussion with Nasťa on little tweaks to make the graph nicer.
And how to get rid of a little hack :)

The date is now displayed more nicely when the mouse is hovering
over the graph, and the Y axis begins at zero.