Track all subpackages with a "python" component in the name
Track all subpackages with `python2`, `python3` or `python` as a
dash-separated component of the name.
For subpackages that don't have a Python dependency (as found by
the existing mechanism), mark them `python2` or `python3`
according to the names. (If it's just `python`, don't mark it
either way.)

This will find a lot of new "misnamed requires".

This should cover `python2-networkx` and similar metapackages, so it
fixes: https://github.com/fedora-python/portingdb/issues/459