Track source RPMs
Source RPMs are added to the list of RPMs, statuses are generated
with source RPMs in mind.
The --qrepo option to the DNF plugin is gone; source/built RPMs are
distinguished by their arch.
The arch is added to RPM entries and SRPMs show up with a wrench
icon in the HTML.