Remove or or embed functions from taskotron-python-versions
Remove the checks for naming policy (is_misnamed and unversioned_requirers).
That part is done; if new packages named python-X appear they
better be python3 (which is not cool according to guidelines,
but not a problem worth tracking).

Remove special-casing NAME_NOTS. This means packages like
python-rpm-generators and python-rpm-macros start showing up as
build dependencies (in python3-only green).

The `have_binaries` function from taskotron-python-versions is embedded.

This is preparation for EOLing taskotron-python-versions.